feat: restructure code.
