+++ rlu/src/rlu.rs
@@ -1,66 +1,22 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-//! # Read-Log-Update (RLU)
-//! ---
-//! This module implements the read log update synchronization mechanism
-//! to enable non-blocking concurrent reads and concurrent writes on
-//! data.
-//!
-//! ## Objective
-//! ---
-//! RLU solves the problem with having multiple concurrent readers being non-block, while having a writer synchronizing
-//! with the reads. RLU still suffers from writer-writer synchronization, eg. two writers with the same memory location
-//! want to update the value.
-//!
-//! ## Algorithm
-//! ---
-//! The main idea is to allow readers non-blocking access to data. RLU employs a global clock (counter)
-//! for (single) versioned updates on objects. On writes, a thread will create a copy of the target
-//! object, locking it before, and conduct any modification to the object inside the log. This ensure
-//! that any modification is hidden from other threads. The writing thread increments the global clock,
-//! so that readers are split into two sets: the readers reading the old state, and readers reading the
-//! new state from the logs of the writing thread.
-//!
-//! The writing thread waits until all readers have concluded their reads, then commits the changes to memory
-//! and update the global clock.
-//!
-//! ## Features
-//! ---
-//! [x] - multiple readers / writers
-//! [x] - lock free
-//!
-//! ## Examples
-//! ---
-//! ```no_run
-//! // no example yet
-//! ```
-//!
-//! # Sources
-//! - [notes](https://chaomai.github.io/2015/2015-09-26-notes-of-rlu/)
-//! - [paper](https://people.csail.mit.edu/amatveev/RLU_SOSP15_paper.pdf)
-//! - [reference impl](https://github.com/rlu-sync/rlu/blob/master/rlu.c)
-//! - [rcu presentation](https://www.cs.unc.edu/~porter/courses/cse506/f11/slides/rcu.pdf)
-
 #![allow(dead_code, unused_variables)]
 
 use log::*;
 use std::{
     collections::HashMap,
-    ops::{Deref, DerefMut, Index, IndexMut},
+    ops::Deref,
     sync::{
         atomic::{AtomicBool, AtomicPtr, AtomicUsize, Ordering},
         Arc,
     },
 };
 
-use crate::BusyBreaker;
+use crate::{Atomic, BusyBreaker, InnerVar, IntoRaw, RLULog, RLUVar, ReadGuard, WriteGuard, WriteGuardInner};
 
 /// Global return type
 pub type Result<T> = core::result::Result<T, TransactionError>;
-
-/// The number of entries inside the write log
-const LOG_ENTRY_SIZE: usize = 32;
 
 #[derive(Debug, thiserror::Error)]
 pub enum TransactionError {
@@ -75,353 +31,6 @@ pub enum TransactionError {
 
     #[error("Operation aborted")]
     Abort,
-}
-
-/// Conversion trait for all types to be heap allocated
-/// and returned as raw mutable pointer
-pub trait IntoRaw: Sized {
-    /// Takes `self`, allocates heap space and retuns it as a raw mutable pointer
-    fn into_raw(self) -> *mut Self {
-        Box::into_raw(Box::new(self))
-    }
-}
-
-impl<T> IntoRaw for T {}
-
-// ----------------- Node Impl
-struct Node<T> {
-    alloc: [Option<T>; LOG_ENTRY_SIZE], // the maximum size of log entries
-    index: AtomicUsize,
-}
-
-impl<T> Node<T>
-where
-    T: Clone,
-{
-    /// Selects the next internal log by incrementing the internal index mod the  number of logs
-    pub fn next_idx(&self) {
-        self.index.fetch_add(1, Ordering::SeqCst);
-    }
-
-    /// Returns the current index  of the log
-    pub fn current_node_index(&self) -> usize {
-        self.index.load(Ordering::SeqCst)
-    }
-
-    pub fn push(&mut self, value: T) {
-        assert!((self.current_node_index() + 1) < self.alloc.len());
-        self.next_idx();
-
-        self.alloc[self.current_node_index()] = Some(value);
-    }
-
-    pub fn clear(&mut self) {
-        self.alloc = match vec![None; LOG_ENTRY_SIZE].try_into() {
-            Ok(array) => array,
-            _ => unreachable!(),
-        };
-        self.index.store(0, Ordering::Release)
-    }
-
-    pub fn len(&self) -> usize {
-        self.alloc.len()
-    }
-
-    pub fn last(&self) -> Option<&T> {
-        self.alloc[self.current_node_index()].as_ref()
-    }
-
-    pub fn iter(&self) -> impl Iterator<Item = &Option<T>> {
-        (&self.alloc).iter()
-    }
-}
-
-impl<T> Default for Node<T>
-where
-    T: Clone,
-{
-    fn default() -> Self {
-        Self {
-            alloc: match vec![None; 32].try_into() {
-                Ok(array) => array,
-                _ => unreachable!(),
-            },
-            index: AtomicUsize::new(0),
-        }
-    }
-}
-
-impl<T> Index<usize> for Node<T>
-where
-    T: Clone,
-{
-    type Output = Option<T>;
-
-    fn index(&self, index: usize) -> &Self::Output {
-        assert!(index < self.alloc.len());
-        &self.alloc[self.current_node_index()]
-    }
-}
-
-impl<T> IndexMut<usize> for Node<T>
-where
-    T: Clone,
-{
-    fn index_mut(&mut self, index: usize) -> &mut Self::Output {
-        assert!(index < self.alloc.len());
-        &mut self.alloc[self.current_node_index()]
-    }
-}
-
-// ----------------- End Node Impl
-
-struct RLULog<T>
-where
-    T: Clone,
-{
-    clear: AtomicUsize,
-    current_log_index: AtomicUsize,
-    logs: [Node<T>; 2], // [Vec<T>; 2],
-}
-
-impl<T> Default for RLULog<T>
-where
-    T: Clone,
-{
-    fn default() -> Self {
-        Self {
-            clear: AtomicUsize::new(usize::MAX),
-            current_log_index: AtomicUsize::new(0),
-            logs: [Node::default(), Node::default()], // unsafe { MaybeUninit::uninit().assume_init() },
-        }
-    }
-}
-
-impl<T> RLULog<T>
-where
-    T: Clone,
-{
-    /// Selects the next internal log by incrementing the internal index mod the  number of logs
-    pub fn next(&self) {
-        let next = (self.current() + 1) % self.logs.len();
-        self.clear.store(self.current(), Ordering::SeqCst);
-        self.current_log_index.store(next, Ordering::SeqCst);
-    }
-
-    /// Returns the current index  of the log
-    pub fn current(&self) -> usize {
-        self.current_log_index.load(Ordering::SeqCst)
-    }
-}
-
-impl<T> Deref for RLULog<T>
-where
-    T: Clone,
-{
-    type Target = Node<T>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.logs[self.current()]
-    }
-}
-
-impl<T> DerefMut for RLULog<T>
-where
-    T: Clone,
-{
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        if let 0..=1 = self.clear.swap(usize::MAX, Ordering::SeqCst) {
-            self.logs[self.current()].clear()
-        }
-
-        &mut self.logs[self.current()]
-    }
-}
-
-/// # Atomic &lt;T&gt;
-///
-/// Wrapper type for [`AtomicPtr`], but with extra heap allocation for the inner type.
-///
-/// ## Example
-/// ```
-/// use stronghold_rlu::rlu::Atomic;
-/// let expected = 1024usize;
-/// let atomic_usize = Atomic::from(expected);
-/// assert_eq!(expected, *atomic_usize);
-/// ```
-pub struct Atomic<T>
-where
-    T: Clone,
-{
-    inner: AtomicPtr<T>,
-}
-
-impl<T> Atomic<T>
-where
-    T: Clone,
-{
-    /// Swaps the inner value and returns the old value.
-    ///
-    /// ## Safety
-    /// This function is unsafe as it tries to dereference a raw pointer which must be allocated
-    /// in accordance to the memory layout of a Box type.
-    pub unsafe fn swap(&self, value: &mut T) -> T {
-        let old = Box::from_raw(self.inner.swap(value, Ordering::SeqCst));
-        *old
-    }
-}
-
-impl<T> Deref for Atomic<T>
-where
-    T: Clone,
-{
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        unsafe { &*self.inner.load(Ordering::SeqCst) }
-    }
-}
-
-impl<T> DerefMut for Atomic<T>
-where
-    T: Clone,
-{
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        unsafe { &mut *self.inner.load(Ordering::SeqCst) }
-    }
-}
-
-impl<T> From<T> for Atomic<T>
-where
-    T: Clone + IntoRaw,
-{
-    fn from(value: T) -> Self {
-        Self {
-            inner: AtomicPtr::new(value.into_raw()),
-        }
-    }
-}
-
-impl<T> Clone for Atomic<T>
-where
-    T: Clone,
-{
-    /// This creates and returns a copy of the pointer to the inner value, not a copy of the value itself
-    fn clone(&self) -> Self {
-        Self {
-            inner: AtomicPtr::new(self.inner.load(Ordering::SeqCst)),
-        }
-    }
-}
-
-pub struct ReadGuard<'a, T>
-where
-    T: Clone,
-{
-    inner: Result<&'a Atomic<T>>,
-    thread: &'a RluContext<T>,
-}
-
-impl<'a, T> ReadGuard<'a, T>
-where
-    T: Clone,
-{
-    pub fn new(inner: Result<&'a Atomic<T>>, thread: &'a RluContext<T>) -> Self {
-        Self { inner, thread }
-    }
-}
-
-impl<'a, T> Drop for ReadGuard<'a, T>
-where
-    T: Clone,
-{
-    fn drop(&mut self) {
-        if let Ok(inner) = self.inner {
-            self.thread.read_unlock(inner)
-        }
-    }
-}
-
-impl<'a, T> Deref for ReadGuard<'a, T>
-where
-    T: Clone,
-{
-    type Target = Result<&'a Atomic<T>>;
-    fn deref(&self) -> &Self::Target {
-        &self.inner
-    }
-}
-
-/// This inner enum of [`WriteGuard`]
-pub enum WriteGuardInner<'a, T>
-where
-    T: Clone,
-{
-    /// a mutable reference
-    Ref(&'a mut T),
-
-    /// a copy, that needs to be written back into the log
-    Copy(InnerVar<T>),
-}
-
-pub struct WriteGuard<'a, T>
-where
-    T: Clone,
-{
-    inner: WriteGuardInner<'a, T>,
-    context: &'a mut RluContext<T>,
-}
-
-impl<'a, T> Deref for WriteGuard<'a, T>
-where
-    T: Clone,
-{
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        match &self.inner {
-            WriteGuardInner::Copy(copy) => match copy {
-                InnerVar::Copy { data, .. } | InnerVar::Original { data, .. } => data,
-            },
-            WriteGuardInner::Ref(reference) => reference,
-        }
-    }
-}
-
-impl<'a, T> DerefMut for WriteGuard<'a, T>
-where
-    T: Clone,
-{
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        match &mut self.inner {
-            WriteGuardInner::Copy(copy) => match copy {
-                InnerVar::Copy { ref mut data, .. } | InnerVar::Original { ref mut data, .. } => data,
-            },
-            WriteGuardInner::Ref(reference) => *reference,
-        }
-    }
-}
-
-impl<'a, T> WriteGuard<'a, T>
-where
-    T: Clone,
-{
-    pub fn new(inner: WriteGuardInner<'a, T>, context: &'a mut RluContext<T>) -> Self {
-        Self { inner, context }
-    }
-}
-
-impl<'a, T> Drop for WriteGuard<'a, T>
-where
-    T: Clone,
-{
-    fn drop(&mut self) {
-        if let WriteGuardInner::Copy(inner) = &self.inner {
-            self.context.log.push(inner.clone())
-        }
-
-        self.context.write_unlock();
-    }
 }
 
 pub struct RLUObject<T>
@@ -465,146 +74,6 @@ where
         Self {
             rlu: self.rlu.clone(),
             var: self.var.clone(),
-        }
-    }
-}
-
-/// # RLUVar &lt;T&gt;
-/// This type represents an rlu managed type. The type is is not being constructed directly
-/// but with the help of [`RLU`].
-///
-/// # Examples
-/// ```no_run
-/// ```
-pub struct RLUVar<T>
-where
-    T: Clone,
-{
-    inner: Arc<AtomicPtr<InnerVar<T>>>,
-}
-
-impl<T> RLUVar<T>
-where
-    T: Clone,
-{
-    /// This function returns the inner value.
-    pub fn get(&self) -> &T {
-        match unsafe { &*self.inner.load(Ordering::SeqCst) } {
-            InnerVar::Copy { data, .. } | InnerVar::Original { data, .. } => data,
-        }
-    }
-}
-
-impl<T> Deref for RLUVar<T>
-where
-    T: Clone,
-{
-    type Target = InnerVar<T>;
-    fn deref(&self) -> &Self::Target {
-        unsafe { &*self.inner.load(Ordering::SeqCst) }
-    }
-}
-
-impl<T> DerefMut for RLUVar<T>
-where
-    T: Clone,
-{
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        unsafe { &mut *self.inner.load(Ordering::SeqCst) }
-    }
-}
-
-impl<T> From<T> for RLUVar<T>
-where
-    T: Clone,
-{
-    fn from(value: T) -> Self {
-        RLUVar {
-            inner: Arc::new(AtomicPtr::new(InnerVar::from(value).into_raw())),
-        }
-    }
-}
-
-impl<T> Clone for RLUVar<T>
-where
-    T: Clone,
-{
-    fn clone(&self) -> Self {
-        Self {
-            inner: self.inner.clone(),
-        }
-    }
-}
-
-pub enum InnerVar<T>
-where
-    T: Clone,
-{
-    Original {
-        locked_thread_id: Option<AtomicUsize>,
-        copy: Option<AtomicPtr<Self>>,
-        data: Atomic<T>,
-        ctrl: Option<RLU<T>>,
-    },
-
-    Copy {
-        locked_thread_id: Option<AtomicUsize>,
-        original: AtomicPtr<Self>,
-        data: Atomic<T>,
-        ctrl: Option<RLU<T>>,
-    },
-}
-
-impl<T> From<T> for InnerVar<T>
-where
-    T: Clone,
-{
-    fn from(value: T) -> Self {
-        Self::Original {
-            data: value.into(),
-            locked_thread_id: None,
-            copy: None,
-            ctrl: None,
-        }
-    }
-}
-
-impl<T> Clone for InnerVar<T>
-where
-    T: Clone,
-{
-    fn clone(&self) -> Self {
-        match self {
-            Self::Copy {
-                ctrl,
-                data,
-                locked_thread_id,
-                original,
-            } => Self::Copy {
-                ctrl: ctrl.clone(),
-                data: data.clone(),
-                locked_thread_id: Some(AtomicUsize::new(match locked_thread_id {
-                    Some(inner) => inner.load(Ordering::SeqCst),
-                    None => 0,
-                })),
-                original: AtomicPtr::new(unsafe { &mut *original.load(Ordering::SeqCst) }),
-            },
-            Self::Original {
-                copy,
-                ctrl,
-                data,
-                locked_thread_id,
-            } => Self::Original {
-                copy: copy
-                    .as_ref()
-                    .map(|inner| AtomicPtr::new(unsafe { &mut *inner.load(Ordering::SeqCst) })),
-                ctrl: ctrl.clone(),
-                data: data.clone(),
-                locked_thread_id: Some(AtomicUsize::new(match locked_thread_id {
-                    Some(inner) => inner.load(Ordering::SeqCst),
-                    None => 0,
-                })),
-            },
         }
     }
 }
@@ -1010,7 +479,7 @@ where
         self.run_count.fetch_add(1, Ordering::SeqCst);
     }
 
-    fn read_unlock(&self, var: &Atomic<T>) {
+    pub(crate) fn read_unlock(&self, var: &Atomic<T>) {
         self.run_count.fetch_add(1, Ordering::SeqCst);
 
         if self.is_writer.load(Ordering::SeqCst) {
@@ -1018,12 +487,16 @@ where
         }
     }
 
-    fn write_lock(&self) {
+    pub(crate) fn write_lock(&self) {
         self.is_writer.store(true, Ordering::SeqCst);
     }
 
-    fn write_unlock(&self) {
+    pub(crate) fn write_unlock(&self) {
         self.is_writer.store(false, Ordering::SeqCst);
+    }
+
+    pub(crate) fn inner_log(&mut self) -> &mut RLULog<InnerVar<T>> {
+        &mut self.log
     }
 
     fn synchronize(&self) {
@@ -1088,82 +561,5 @@ where
     /// Swaps the logs internally
     fn swap_logs(&self) {
         self.log.next();
-    }
-}
-
-#[cfg(test)]
-mod tests {
-
-    use super::Atomic;
-    use crate::rlu::RLULog;
-    use rand_utils::random::{string, usize};
-
-    fn rand_string() -> String {
-        string(255)
-    }
-
-    #[inline(always)]
-    fn rand_usize() -> usize {
-        usize(usize::MAX)
-    }
-
-    #[test]
-    fn test_rlu_log() {
-        let mut log = RLULog::<usize>::default();
-
-        // 0
-        assert_eq!(log.current(), 0);
-        log.push(1);
-        log.push(1);
-        assert_eq!(log.current_node_index(), 2);
-
-        // 1
-        log.next();
-        assert_eq!(log.current(), 1);
-        log.push(1);
-        log.push(1);
-        log.push(1);
-        assert_eq!(log.current_node_index(), 3);
-
-        // 0
-        log.next();
-        assert_eq!(log.current(), 0);
-        log.push(1);
-        log.push(1);
-        assert_eq!(log.current_node_index(), 2);
-
-        // 1
-        log.next();
-        assert_eq!(log.current(), 1);
-        log.push(1);
-        log.push(1);
-        log.push(1);
-        log.push(1);
-        log.push(1);
-        assert_eq!(log.current_node_index(), 5);
-
-        // 0
-        log.next();
-        assert_eq!(log.current(), 0);
-        log.push(1);
-        log.push(1);
-        assert_eq!(log.current_node_index(), 2);
-    }
-
-    #[test]
-    fn test_atomic_type() {
-        let num_runs = 1000;
-
-        for _ in 0..num_runs {
-            let expected = rand_string();
-            let mut expected_mod = expected.clone();
-            expected_mod.push_str("_modified");
-
-            let atomic_string = Atomic::from(expected.clone());
-            assert_eq!(expected, *atomic_string);
-
-            unsafe { atomic_string.swap(&mut expected_mod) };
-            assert_eq!(expected_mod, *atomic_string);
-        }
     }
 }
